Walker Algorithm Implemented
Using the class 'walkerbot', this function implements a walker algorithm, that is, the robot moves straight until it reaches an obstacle. Then it rotates about its position until it is free of obstacle. Then the robot again moves straight. The obstacle is detected using the laser scan data from the turtlebot.