Simulation World Created
A simulation world has been created to simulate the walker algorithm. In this world, there are two humans and two dumpsters within walled confines. The walls make sure that the robot keeps on running. However, the obstacles in between are created to show the difference when the maximum obstacle clearance is changed.